requirements.txt
numpy==1.26.2
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
argumentparser==1.2.1